Fix install packages

+++ run_once_install-packages.sh
@@ -6,5 +6,5 @@ command -v op >/dev/null || brew install --cask 1password/tap/1password-cli
 command -v fish >/dev/null || brew install fish
 command -v starship >/dev/null || brew install starship
 brew install font-symbols-only-nerd-font
-brew install --cask iterm2
-defaults write com.googlecode.iterm2 PrefsCustomFolder -string 
+test -e /Applications/iTerm2.app || brew install --cask iterm2
+defaults write com.googlecode.iterm2 PrefsCustomFolder -string $HOME/.local/share/chezmoi/iterm2
